@@ -1,43 +1,34 @@
 {
-  "name": "quacker-bot",
-  "version": "1.0.0",
-  "description": "Discord bot for the Quackers server and friends",
-  "main": "out/index.js",
-  "scripts": {
-    "db": "mongod --auth --port=80 --dbpath=$(pwd)/.data",
-    "start": "node out/index.js",
-    "deploy": "node out/deploy-commands.js",
-    "deploy-global": "node out/deploy-commands-global.js",
-    "watch": "tsc -p tsconfig.json -w",
-    "build": "tsc -p tsconfig.json",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/exerro/quacker-bot.git"
-  },
-  "keywords": [
-    "discord",
-    "bot",
-    "quackers"
-  ],
-  "author": "exerro",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/exerro/quacker-bot/issues"
-  },
-  "homepage": "https://github.com/exerro/quacker-bot#readme",
-  "dependencies": {
-    "@discordjs/builders": "^0.8.2",
-    "@discordjs/rest": "^0.1.0-canary.0",
-    "@types/mongodb": "^3.6.20",
-    "discord-api-types": "^0.24.0",
-    "discord.js": "^13.3.1",
-    "dotenv": "^10.0.0",
-    "mongodb": "^3.7.3"
-  },
-  "devDependencies": {
-    "@types/node": "^15.14.9",
-    "typescript": "^4.4.4"
-  }
+	"name": "quacker-bot",
+	"version": "1.1.0",
+	"description": "Discord bot for the Quackers server and friends",
+	"main": "out/index.js",
+	"author": "exerro",
+	"license": "MIT",
+	"keywords": [
+	  "discord",
+	  "bot",
+	  "quackers"
+	],
+	"repository": {
+	  "type": "git",
+	  "url": "git+https://github.com/exerro/quacker-bot.git"
+	},
+	"bugs": {
+	  "url": "https://github.com/exerro/quacker-bot/issues"
+	},
+	"packageManager": "yarn@3.0.2",
+	"homepage": "https://github.com/exerro/quacker-bot#readme",
+	"dependencies": {
+	  "@discordjs/builders": "^0.8.2",
+	  "@discordjs/rest": "^0.1.1-canary.0",
+	  "discord.js": "^13.3.1",
+	  "mongodb": "^3.7.3"
+	},
+	"devDependencies": {
+		"@types/node": "^15.14.9",
+		"@types/mongodb": "^3.6.20",
+		"discord-api-types": "^0.24.0",
+		"typescript": "^4.5.2"
+	}
 }
