@@ -1,6 +1,8 @@
-import { Snowflake } from "discord.js"
 
-export type Quote = {
+export type Quote = QuoteV1_1
+
+export type QuoteV1_1 = {
+	quoteVersion: '1.1',
 	id: number,
 	quoter: string,
 	multiQuoteParts: {
