@@ -1,5 +1,4 @@
-import { CacheType, CommandInteraction, User } from "discord.js";
-import { Db } from "mongodb";
+import { User } from "discord.js";
 import { DatabaseInterface } from "./database";
 
 export type State = {
