@@ -1,25 +1,45 @@
 
 import { connect as db_connect } from "./database"
 import { Client, Intents } from "discord.js"
-import { config } from "dotenv"
 import { Command, loadCommands } from "./commands/commands"
 import { State } from "./State"
 import Collection from "@discordjs/collection"
-
-config()
+import { REST } from '@discordjs/rest'
+import { Routes } from 'discord-api-types/v9'
 
 console.log('Starting QuackerBot')
 
-db_connect((db) => {
-	let client = new Client({ intents: [Intents.FLAGS.GUILDS] })
-	let commands = new Collection<string, Command>()
-	let state: State = { database: db, currentMultiQuotes: new Map() }
+async function deploy() {
+	const slashCommands = (await loadCommands()).map(c => c.command.toJSON())
+	const rest = new REST({ version: '9' }).setToken(process.env.DISCORD_TOKEN || "")
 
-	loadCommands().then(loadedCommand => {
-		for (const command of loadedCommand) {
-			commands.set(command.command.name, command)
-		}
-	})
+	let route: `/${string}`
+
+	if (process.env["DEBUG"]) {
+		console.log("Deploying in debug mode")
+		route = Routes.applicationGuildCommands(process.env.DISCORD_CLIENT_ID || "", process.env.DISCORD_GUILD_ID || "")
+	}
+	else {
+		console.log("Deploying in production mode")
+		route = Routes.applicationCommands(process.env.DISCORD_CLIENT_ID || "")
+	}
+
+	await rest.put(route, { body: slashCommands })
+
+	console.log(`Successfully registered ${slashCommands.length} application commands.`)
+}
+
+db_connect(async (db) => {
+	await deploy()
+
+	let state: State = { database: db, currentMultiQuotes: new Map() }
+	const client = new Client({ intents: [Intents.FLAGS.GUILDS] })
+	const commands = new Collection<string, Command>()
+	const loadedCommands = await loadCommands()
+
+	for (const command of loadedCommands) {
+		commands.set(command.command.name, command)
+	}
 
 	client.once('ready', () => {
 		console.log('QuackerBot is ready!')
@@ -48,4 +68,4 @@ db_connect((db) => {
 	})
 
 	client.login(process.env.DISCORD_TOKEN)
-})
+}).catch(console.error)
